+++ src/js/base.js
@@ -11,24 +11,29 @@ window.addEventListener("load", () => {
         const app = new App(mainFolder.replaceAll("\\", "/"));
 
         window.addEventListener("beforeunload", (e) => {
-            const savesFile = path.join(app.mainFolder, "data", "saves.json");
-            const saveData = {
-                volume: app.elements.footer.volume.slider.value / 100,
-                loop: app.settings.loop,
-                random: app.settings.random,
-                window: {
-                    x: Math.max(app.window.x, 0),
-                    y: Math.max(app.window.y, 0),
-                    w: app.window.w,
-                    h: app.window.h,
-                    f: app.window.f,
-                },
-                playlists: [],
-            };
-            const strSaveData = JSON.stringify(saveData, null, 2);
-            fs.writeFile(savesFile, strSaveData, (err) => {
-                if (err) console.error("Error writing json:", err);
-            });
+            const settingsFile = path.join(app.mainFolder, "data", "settings.json");
+            fsp.readFile(settingsFile, "utf-8").then((data) => {
+                const jsonData = JSON.parse(data);
+
+                const saveData = {
+                    volume: app.elements.footer.volume.slider.value / 100,
+                    loop: app.settings.loop,
+                    random: app.settings.random,
+                    window: {
+                        x: Math.max(app.window.x, 0),
+                        y: Math.max(app.window.y, 0),
+                        w: app.window.w,
+                        h: app.window.h,
+                        f: app.window.f,
+                    },
+                    playlists: [],
+                    colors: jsonData.colors,
+                };
+                const strSaveData = JSON.stringify(saveData, null, 2);
+                fs.writeFile(settingsFile, strSaveData, (err) => {
+                    if (err) console.error("Error writing json:", err);
+                });
+            }).catch((readErr) => console.error("Error => can't read songs.json:" + readErr));
         });
     });
 });